feat: refactor state and API machines
